% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{bitJowoncni 패키지가 부착될 때 실행되는 함수입니다.}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{패키지의 라이브러리 경로}

\item{pkgname}{패키지의 이름}
}
\description{
이 함수는 패키지 사용자에게 변수의 명명 규칙에 관한 메시지를 제공합니다.
}
\keyword{internal}
