% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_st_news_sav.R
\name{load_st_news_sav}
\alias{load_st_news_sav}
\title{Load a .sav file from the st_news/sav directory of the bitJowoncni package}
\usage{
load_st_news_sav(filename)
}
\arguments{
\item{filename}{The name of the .sav file you want to load from the st_news/sav directory.}
}
\value{
A data frame containing the loaded data.
}
\description{
This function provides a convenient way to load .sav files that are stored
in the st_news/sav directory of the bitJowoncni package.
}
\examples{
\dontrun{
data <- load_st_news_sav("example_file.sav")
head(data)
}
}
