% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runJowoncni.R
\name{runJowoncni}
\alias{runJowoncni}
\title{Launch the Shiny App from bitJowoncni package}
\usage{
runJowoncni()
}
\description{
This function runs the Shiny app included in the bitJowoncni package.
}
\examples{
\dontrun{
runJowoncni()
}
}
